[refactor/#24] API 호출 함수의 타입을 모듈화된 위치로 이동
